+++ sw.js
@@ -83,6 +83,7 @@ const PRECACHE_URLS = [
   './js/engine/responder-entity.js',
   './js/engine/responder-overrides.js',
   './js/engine/responder-recall.js',
+  './js/engine/responder-lifefacts.js',
   './js/engine/responder-profile.js',
   './js/engine/responder-promise.js',
   './js/engine/responder-exercises.js',